requirements.txt
Flask
folium
geopy
requests
gunicorn
python-dotenv
openai
